chore(hooks): payment usehooks organized
