+++ src/components/membership-components/AddMemberDialog.tsx
@@ -33,7 +33,7 @@ import { getAuthenticatedApi } from '@/lib/api-client'
 import { useAuthSession } from '@/lib/auth-session'
 import { AddBillingDialog } from '@/components/membership-components/AddBillingForm'
 import { useAddMemberDialogData } from '@/hooks/membership/useAddMemberDialogData'
-import { useBillingActions } from '@/hooks/useBillingActions'
+import { useBillingActions } from '@/hooks/billing/useBillingActions'
 import { memberQueryKeys } from '@/lib/QueryKeys'
