+++ src/components/membership-components/MemberDetailsDialog.tsx
@@ -28,7 +28,7 @@ import { useAuthSession } from '@/lib/auth-session'
 import { AddBillingDialog } from './AddBillingForm'
 import { useMemberDetailsDialogData } from '@/hooks/membership/useMemberDetailsDialogData'
 import { isAdminToken } from '@/lib/auth-permissions'
-import { useBillingActions } from '@/hooks/useBillingActions'
+import { useBillingActions } from '@/hooks/billing/useBillingActions'
 import { memberQueryKeys } from '@/lib/QueryKeys'
