+++ src/lib/QueryKeys.ts
@@ -25,3 +25,5 @@
   export const userQueryKeys = {
     currentUser: 'current-user',
   }
+
+
