+++ src/routes/dashboard/billing/payment-history.tsx
@@ -6,7 +6,9 @@ import { useQuery } from '@tanstack/react-query'
 import { format } from 'date-fns'
 import { AlertTriangle, CalendarIcon, CheckCircle2, Clock, Receipt, Search, XCircle } from 'lucide-react'
 
-import { usePayment, usePaymentMethods, usePayments, useInvoices } from '@/hooks/usePaymentHistory'
+import { usePayments, usePayment } from '@/hooks/billing/usePayments'
+import { usePaymentMethods } from '@/hooks/billing/usePaymentMethods'
+import { useInvoices } from '@/hooks/billing/useInvoices'
 import { getAuthenticatedApi } from '@/lib/api-client'
 import { MemberApi } from '@/api/generated/apis/MemberApi'
 import { apiResponseListMemberTableSchema } from '@/types/membership/memberSchemas'
