+++ src/routes/dashboard/marketing/membership.tsx
@@ -63,6 +63,7 @@ function MembershipRoute() {
 
   return (
     <div className="space-y-6">
+      <h1 className="text-2xl font-bold text-primary">Membership Management</h1>
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
         <MembersTable
           members={members}
